Move logging into the internal package.
